requirements.txt
pyTelegramBotAPI
python-dotenv
psycopg2-binary
beautifulsoup4
requests
pandas